+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 * Migrate to null safety
 
+## 0.0.6
+Thanks to nerder for the following feature
+
+* Added the option to set the title of the native picker for Android
+* Added title parameter for MailAppPickerDialog
+
 ## 0.0.5
 
 * Fixed MailApp name being null when building on Android when minifyEnabled is set to true